+++ src/Components/Frontoffice/Map/MapView.css
@@ -1,0 +1,4 @@
+.leaflet-container {
+    width: 100vw;
+    height: 100vh;
+}
